+++ local_io/local_io.cpp
@@ -26,7 +26,7 @@ class DefaultCurAttrProvider final : public wwiv::local_io::curatr_provider {
 public:
   DefaultCurAttrProvider() = default;
   [[nodiscard]] uint8_t curatr() const noexcept override { return a_; }
-  void curatr(uint8_t n) override { a_ = n; }
+  void curatr(int n) override { a_ = static_cast<uint8_t>(n); }
 
 private:
   uint8_t a_{7};
@@ -42,7 +42,7 @@ void LocalIO::set_curatr_provider(wwiv::local_io::curatr_provider* p) { curatr_ 
 wwiv::local_io::curatr_provider* LocalIO::curatr_provider() { return curatr_; }
 
 uint8_t LocalIO::curatr() const { return curatr_->curatr(); }
-void LocalIO::curatr(uint8_t c) { curatr_->curatr(c); }
+void LocalIO::curatr(int c) { curatr_->curatr(c); }
 
 int LocalIO::EditLine(std::string& s, int len, AllowedKeys allowed_keys,
                     const std::string& allowed_set_chars) {
